# Increase parsing and comparison performance