--- a/yaml.go
+++ b/yaml.go
@@ -26,12 +26,12 @@ func (v *Version) UnmarshalYAML(node *yaml.Node) error {
 		return err
 	}
 
+	v.raw = parsed.raw
 	v.major = parsed.major
 	v.minor = parsed.minor
 	v.patch = parsed.patch
-	v.prerelases = parsed.prerelases
-	v.numericPrereleases = parsed.numericPrereleases
-	v.builds = parsed.builds
+	v.prerelease = parsed.prerelease
+	v.build = parsed.build
 	return nil
 }