--- a/yaml.go
+++ b/yaml.go
@@ -27,6 +27,7 @@ func (v *Version) UnmarshalYAML(node *yaml.Node) error {
 	}
 
 	v.raw = parsed.raw
+	v.bytes = []byte(v.raw)
 	v.major = parsed.major
 	v.minor = parsed.minor
 	v.patch = parsed.patch