… binary_test.go
… json_test.go
… parser_test.go
… version_test.go
… yaml_test.go