This patch does a lot of performance improvement by inlining hot paths and by reducing memory pressure removing unneeded allocation.

Before:
```
BenchmarkVersionParser-12         188611     7715 ns/op     8557 B/op    51 allocs/op
BenchmarkVersionComparator-12      74793    17347 ns/op        0 B/op     0 allocs/op
```

After:
```
BenchmarkVersionParser-12         479626     3719 ns/op      616 B/op    51 allocs/op (+154% !!)
BenchmarkVersionComparator-12     101772    11720 ns/op        0 B/op     0 allocs/op (+36%)
```